requirements.txt
eel==0.16.0
pyinstaller==6.3.0
docxtpl==0.16.7